feat: moved more functions to node modules
