+++ gulpfile.js
@@ -1,11 +1,11 @@
 const gulp = require('gulp');
 const postcss = require('gulp-postcss');
-const easysprite = require('./index.js');
 const rename = require('gulp-rename');
 const eslint = require('gulp-eslint');
 const mocha = require('gulp-mocha');
+const easysprite = require('./index.js');
 
-gulp.task('project:basic', function(done) {
+gulp.task('project:basic', (done) => {
   gulp
     .src('./test/basic/input.css')
     .pipe(
@@ -22,19 +22,19 @@ gulp.task('project:basic', function(done) {
   done();
 });
 
-gulp.task('linting', function() {
+gulp.task('linting', () => {
   return gulp.src('./index.js').pipe(eslint()); // hint (optional)
 });
 
-gulp.task('runtest', function() {
+gulp.task('runtest', () => {
   return gulp.src('test/basic.js', { read: false }).pipe(mocha());
 });
 
-gulp.task('test', gulp.series('project:basic', 'linting'), function() {
+gulp.task('test', gulp.series('project:basic', 'linting'), () => {
   return gulp.src('test/*.js', { read: false }).pipe(mocha());
 });
 
-gulp.task('watch', function(done) {
+gulp.task('watch', (done) => {
   gulp.watch(['test/basic/input.css'], ['project:basic']);
 
   done();
