+++ lib/apply-group-by.js
@@ -4,15 +4,15 @@ const Q = require('q');
 /**
  * TODO: Group by.
  *
- * @param {Array} imagesToGroup - Array of image objects.
+ * @param {Array} imageCollection - Array of image objects.
  * @param {object} opts - Options passed to the plugin.
  * @returns {Promise}
  */
-function applyGroupBy(imagesToGroup, opts) {
+function applyGroupBy(imageCollection, opts) {
   return Q.Promise((resolve, reject) => {
     async.reduce(
       opts.groupBy,
-      imagesToGroup,
+      imageCollection,
       (images, group, next) => {
         async.map(
           images,
@@ -33,6 +33,7 @@ function applyGroupBy(imagesToGroup, opts) {
         if (err) {
           return reject(err);
         }
+
         return resolve([images, opts]);
       }
     );
