+++ lib/collect-images.js
@@ -12,7 +12,7 @@ const {
 } = require('./image-urls');
 
 /**
- * TODO: Collect images...
+ * Collect all background images into an array.
  *
  * @param {object} css - Object with CSS results.
  * @param {object} [opts] - Options passed to the plugin.
@@ -22,11 +22,14 @@ function collectImages(css, opts) {
   const images = [];
   const stylesheetPath =
     opts.stylesheetPath || path.dirname(css.source.input.file);
+
   if (!stylesheetPath) {
     throw new Error(
       'Stylesheets path is undefined, please use option stylesheetPath!'
     );
   }
+
+  // Loop through all CSS rules.
   css.walkRules((rule) => {
     const image = {
       path: null,
@@ -36,9 +39,11 @@ function collectImages(css, opts) {
       groups: [],
       token: '',
     };
+
     if (hasImageInRule(rule.toString())) {
       image.url = getImageUrl(rule.toString());
       const imageUrl = url.parse(image.url);
+
       // only locals, hashed paths
       if (
         imageUrl.host ||
@@ -48,31 +53,39 @@ function collectImages(css, opts) {
       ) {
         return;
       }
+
       image.hash = imageUrl.hash.replace('#', '');
       image.groups = [image.hash];
-      // Perform search for retina
+
+      // Check if the image is retina.
       if (isRetinaImage(image.url)) {
         image.ratio = getRetinaRatio(image.url);
       }
+
       // Get the path to the image.
       image.path = resolveImageUrl(image, opts);
-      // file exists
+
+      // Check that the file exists.
       if (!fs.existsSync(image.path)) {
         log(
           'Easysprites:',
           ansi.red(image.path),
-          'file unreachable or not exists'
+          'file unreachable or does not exists'
         );
-        // Remove hash from link.
-        lodash.each(rule.nodes, (node) => {
-          const ruleNode = node;
-          ruleNode.value = ruleNode.value.replace(`#${image.hash}`, '');
+
+        // Cleanup background image URL's by removing hash suffix.
+        rule.nodes.forEach((node) => {
+          const nodeRule = node;
+          nodeRule.value = node.value.replace(`#${image.hash}`, '');
         });
+
         return;
       }
+
       images.push(image);
     }
   });
+
   return lodash.uniqWith(images, lodash.isEqual);
 }
