+++ lib/get-background-values.js
@@ -7,6 +7,7 @@
 function getBackgroundSize(image) {
   const x = image.width / image.ratio;
   const y = image.height / image.ratio;
+
   return `${x}px ${y}px`;
 }
 
@@ -34,20 +35,20 @@ function getBackgroundPosition(image) {
 }
 
 /**
- * Extract the background color from declaration.
+ * Extract the background color from the CSS declaration.
  *
- * @param {object} decl - The process CSS declaration.
+ * @param {string} backgroundValue - The background CSS property value.
  * @returns {string|null}
  */
-function getBackgroundColor(decl) {
-  const regexes = ['(#([0-9a-f]{3}){1,2})', 'rgba?\\([^\\)]+\\)'];
+function getBackgroundColor(backgroundValue) {
+  const colorPatterns = ['(#([0-9a-f]{3}){1,2})', '(rgb|hsl)a?\\([^\\)]+\\)'];
   let matches = null;
 
-  regexes.forEach((regexPattern) => {
+  colorPatterns.forEach((regexPattern) => {
     const regex = new RegExp(regexPattern, 'gi');
 
-    if (regex.test(decl.value)) {
-      matches = decl.value.match(regex);
+    if (regex.test(backgroundValue)) {
+      [matches] = backgroundValue.match(regex);
     }
   });
