+++ lib/log.js
@@ -1,7 +1,7 @@
 const fancyLog = require('fancy-log');
 
 /**
- * Custom log function to style status messages.
+ * Custom console log function to output styled status messages.
  *
  * @param {object} message - A message to print to the console with formatting.
  *
