+++ lib/tokens.js
@@ -41,9 +41,8 @@ function setTokens(images, opts, css) {
         });
 
         if (image) {
-          // We remove these declarations since
-          // our plugin will insert them when
-          // they are necessary.
+          // We remove these declarations since our plugin will insert them
+          // when they are necessary.
           rule.walkDecls(
             /^background-(repeat|size|position)$/,
             (backgroundDecl) => {
@@ -52,12 +51,12 @@ function setTokens(images, opts, css) {
           );
 
           if (decl.prop === BACKGROUND) {
-            color = getBackgroundColor(decl);
+            color = getBackgroundColor(decl.value);
 
-            // Extract color to background-color propery
-            if (color && color.length === 1) {
+            // Extract the color for the background-color property.
+            if (color) {
               rule.prop = 'background-color';
-              [rule.value] = color;
+              rule.value = color;
               rule.before = ' ';
             }
           }
