+++ test/basic.js
@@ -8,6 +8,12 @@ const fixture = new StdOutFixture();
 // Keep track of writes so we can check them later..
 const logCapture = [];
 
+const testOptions = {
+  imagePath: './test/basic',
+  stylesheetPath: './test/basic', // need here cause of inline call
+  spritePath: './test/basic/sprites',
+};
+
 /**
  * Saves the console output so previous messages can be retrieved.
  *
@@ -37,6 +43,7 @@ const assert = (input, output, opts, done) => {
     .process(input, { from: undefined })
     .then((result) => {
       expect(result.css).to.eql(output);
+      /* eslint-disable no-unused-expressions */
       expect(result.warnings()).to.be.empty;
       done();
     })
@@ -53,6 +60,7 @@ const assertNotCached = (input, output, opts, done) => {
     .process(input, { from: undefined })
     .then((result) => {
       expect(result.css).to.eql(output);
+      /* eslint-disable no-unused-expressions */
       expect(result.warnings()).to.be.empty;
 
       // Stop stdout capture.
@@ -93,7 +101,7 @@ const assertCached = (input, output, opts, done) => {
     });
 };
 
-describe('postcss-easysprites', () => {
+describe('postcss-easysprites: basic', () => {
   it('Relative images test', (done) => {
     assert(
       'a { background:url("images/arrow-next.png#elements"); }',
@@ -110,9 +118,17 @@ describe('postcss-easysprites', () => {
     assert(
       'a { background: url("/images/arrow-next.png#elements"); }',
       'a { background-image: url(sprites/elements.png); background-position: 0 0; }',
+      JSON.parse(JSON.stringify(testOptions)),
+      done
+    );
+  });
+
+  it('Background image property test', (done) => {
+    assert(
+      'a { background-image: url("images/arrow-next.png#elements"); }',
+      'a { background-image: url(sprites/elements.png); background-position: 0 0; }',
       {
-        imagePath: './test/basic',
-        stylesheetPath: './test/basic', // need here cause of inline call
+        stylesheetPath: './test/basic',
         spritePath: './test/basic/sprites',
       },
       done
@@ -123,11 +139,7 @@ describe('postcss-easysprites', () => {
     assert(
       'a { background: url("/images/arrow-next@2x.png#elements"); }',
       'a { background-image: url(sprites/elements@2x.png); background-position: 0 0; background-size: 28px 27px; }',
-      {
-        imagePath: './test/basic',
-        stylesheetPath: './test/basic', // need here cause of inline call
-        spritePath: './test/basic/sprites',
-      },
+      JSON.parse(JSON.stringify(testOptions)),
       done
     );
   });
@@ -136,37 +148,65 @@ describe('postcss-easysprites', () => {
     assert(
       'a { background: url("/images/image-not-exists.png#elements"); }',
       'a { background: url("/images/image-not-exists.png"); }',
-      {
-        imagePath: './test/basic',
-        stylesheetPath: './test/basic', // need here cause of inline call
-        spritePath: './test/basic/sprites',
-      },
+      JSON.parse(JSON.stringify(testOptions)),
       done
     );
   });
+});
 
+describe('postcss-easysprites: colors', () => {
   it('Hex background color test', (done) => {
     assert(
       'a { background: #000000 url("/images/arrow-next.png#elements"); }',
-      'a { background-image: url(sprites/elements.png); background-position: 0 0; }',
-      {
-        imagePath: './test/basic',
-        stylesheetPath: './test/basic', // need here cause of inline call
-        spritePath: './test/basic/sprites',
-      },
+      'a { background-image: url(sprites/elements.png); background-position: 0 0; background-color: #000000; }',
+      JSON.parse(JSON.stringify(testOptions)),
       done
     );
   });
 
+  it('RGB background color test', (done) => {
+    assert(
+      'a { background: rgb(0, 0, 0) url("/images/arrow-next.png#elements"); }',
+      'a { background-image: url(sprites/elements.png); background-position: 0 0; background-color: rgb(0, 0, 0); }',
+      JSON.parse(JSON.stringify(testOptions)),
+      done
+    );
+  });
+
+  it('RGBa background color test', (done) => {
+    assert(
+      'a { background: rgba(0, 0, 0, 1) url("/images/arrow-next.png#elements"); }',
+      'a { background-image: url(sprites/elements.png); background-position: 0 0; background-color: rgba(0, 0, 0, 1); }',
+      JSON.parse(JSON.stringify(testOptions)),
+      done
+    );
+  });
+
+  it('HSL background color test', (done) => {
+    assert(
+      'a { background: hsl(0,100%, 50%) url("/images/arrow-next.png#elements"); }',
+      'a { background-image: url(sprites/elements.png); background-position: 0 0; background-color: hsl(0,100%, 50%); }',
+      JSON.parse(JSON.stringify(testOptions)),
+      done
+    );
+  });
+
+  it('HSLa background color test', (done) => {
+    assert(
+      'a { background: hsla(0,100%, 50%, 1) url("/images/arrow-next.png#elements"); }',
+      'a { background-image: url(sprites/elements.png); background-position: 0 0; background-color: hsla(0,100%, 50%, 1); }',
+      JSON.parse(JSON.stringify(testOptions)),
+      done
+    );
+  });
+});
+
+describe('postcss-easysprites: caching', () => {
   it('Assert sprite not cached', (done) => {
     assertNotCached(
       'a { background: url("images/arrow-next_hover.png#elements"); }',
       'a { background-image: url(sprites/elements.png); background-position: 0 0; }',
-      {
-        imagePath: './test/basic',
-        stylesheetPath: './test/basic', // need here cause of inline call
-        spritePath: './test/basic/sprites',
-      },
+      JSON.parse(JSON.stringify(testOptions)),
       done
     );
   });
@@ -175,11 +215,7 @@ describe('postcss-easysprites', () => {
     assertCached(
       'a { background: url("images/arrow-next_hover.png#elements"); }',
       'a { background-image: url(sprites/elements.png); background-position: 0 0; }',
-      {
-        imagePath: './test/basic',
-        stylesheetPath: './test/basic', // need here cause of inline call
-        spritePath: './test/basic/sprites',
-      },
+      JSON.parse(JSON.stringify(testOptions)),
       done
     );
   });
